added api key UI

+++ src/server/controllers/api-key.mjs
@@ -8,11 +8,27 @@ import {
 
 import {isAuthenticated} from '../middleware/authentication.mjs';
 
+const {AWS_ACCESS_KEY_ID} = process.env;
+
 const client = new APIGatewayClient({region: 'us-east-2'});
 
 const ApiKeyRouter = express.Router();
 
 ApiKeyRouter.all('*', isAuthenticated);
+
+ApiKeyRouter.get('/', async (req, res, next) => {
+
+    try {
+        const ApiKey = mongoose.model('ApiKey');
+
+        res.json(await ApiKey.find({
+            user: req.user._id,
+            organization: req.user.activeOrganizationMembership.organization._id
+        }));
+    } catch (e) {
+        next(e);
+    }
+});
 
 ApiKeyRouter.post('/', async (req, res, next) => {
 
@@ -25,21 +41,31 @@ ApiKeyRouter.post('/', async (req, res, next) => {
             user: req.user._id,
             organization: req.user.activeOrganizationMembership.organization._id
         });
-        const awsApiKey = await client.send(new CreateApiKeyCommand({
-            enabled: true,
-            name: `${dioptraApiKey._id} (as: ${req.user.username} | ${activeOrganization.name})`,
-            tags: {
-                dioptraUserId, dioptraOrganizationId,
-                dioptraApiKeyId: dioptraApiKey._id
-            }
-        }));
 
-        await client.send(new CreateUsagePlanKeyCommand({
-            keyId: awsApiKey.id,
-            keyType: 'API_KEY',
-            // TODO: Change this depending on what plan the organization is on when people are paying us.
-            usagePlanId: 'emtm28' // 'Demo Plan'.
-        }));
+        let awsApiKey = {
+            id: '__placeholder_api_key_id__',
+            value: '__placeholder_api_key_value__'
+        };
+
+        if (AWS_ACCESS_KEY_ID) {
+
+            awsApiKey = await client.send(new CreateApiKeyCommand({
+                enabled: true,
+                name: `${dioptraApiKey._id} (as: ${req.user.username} | ${activeOrganization.name})`,
+                tags: {
+                    dioptraUserId, dioptraOrganizationId,
+                    dioptraApiKeyId: dioptraApiKey._id
+                }
+            }));
+
+            await client.send(new CreateUsagePlanKeyCommand({
+                keyId: awsApiKey.id,
+                keyType: 'API_KEY',
+                // TODO: Change this depending on what plan the organization is on when people are paying us.
+                usagePlanId: 'emtm28' // 'Demo Plan'.
+            }));
+        }
+
 
         dioptraApiKey.awsApiKeyId = awsApiKey.id;
         dioptraApiKey.awsApiKey = awsApiKey.value;
@@ -59,9 +85,11 @@ ApiKeyRouter.delete('/:_id', async (req, res, next) => {
             user: req.user._id
         });
 
-        await client.send(new DeleteApiKeyCommand({
-            apiKey: dioptraApiKey.awsApiKeyId
-        }));
+        if (AWS_ACCESS_KEY_ID) {
+            await client.send(new DeleteApiKeyCommand({
+                apiKey: dioptraApiKey.awsApiKeyId
+            }));
+        }
 
         res.json(dioptraApiKey);
     } catch (e) {
